+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter as Router } from 'react-router-dom';
 import Header from '@widgets/ui/Header.tsx';
 import Footer from '@widgets/ui/Footer.tsx';
-import PostsManagerPage from '@pages/PostsManagerPage.tsx';
+import Home from '@/pages/Home';
 
 const App = () => {
   return (
@@ -9,7 +9,7 @@ const App = () => {
       <div className="flex flex-col min-h-screen">
         <Header />
         <main className="flex-grow container mx-auto px-4 py-8">
-          <PostsManagerPage />
+          <Home />
         </main>
         <Footer />
       </div>
